Start refactoring mock task package
