+++ tests/v2/04_runner/fractal_tasks_core_mock/src/fractal_tasks_mock/calculate_registration_compute.py
@@ -1,0 +1,30 @@
+from pathlib import Path
+
+from pydantic import BaseModel
+from pydantic.decorator import validate_arguments
+
+
+class InitArgsRegistration(BaseModel):
+    ref_path: str
+
+
+@validate_arguments
+def calculate_registration_compute(
+    *,
+    path: str,
+    init_args: InitArgsRegistration,
+) -> None:
+
+    ref_path = init_args.ref_path
+    print("[calculate_registration_compute] START")
+    print(f"[calculate_registration_compute] {path=}")
+    print(f"[calculate_registration_compute] {ref_path=}")
+
+    table_path = Path(path) / "registration_table"
+    print(
+        f"[calculate_registration_compute] Writing to {table_path.as_posix()}"
+    )
+
+    with table_path.open("w") as f:
+        f.write("Calculate registration for\n" f"{path=}\n" f"{ref_path=}\n")
+    print("[calculate_registration_compute] END")
