+++ tests/v2/04_runner/fractal_tasks_core_mock/src/fractal_tasks_mock/new_ome_zarr.py
@@ -1,22 +1,18 @@
-import shutil
 from pathlib import Path
 
-from pydantic import BaseModel
 from pydantic.decorator import validate_arguments
 
 from .utils import _extract_common_root
-from fractal_server.app.runner.v2.models import DictStrAny
 
 
 @validate_arguments
 def new_ome_zarr(
     *,
-    # Standard arguments
     paths: list[str],
     zarr_dir: str,
-    # Non-standard arguments
     suffix: str = "new",
 ) -> dict:
+
     dict_shared = _extract_common_root(paths)
     old_plate = dict_shared.get("shared_plate")
 
@@ -51,37 +47,4 @@ def new_ome_zarr(
     # Prepare output metadata
     out = dict(parallelization_list=parallelization_list)
     print("[new_ome_zarr] END")
-    return out
-
-
-class InitArgsMIP(BaseModel):
-    new_path: str
-    new_plate: str  # FIXME: remove this
-
-
-@validate_arguments
-def maximum_intensity_projection(
-    *,
-    path: str,
-    init_args: InitArgsMIP,
-) -> DictStrAny:
-    new_path = init_args.new_path
-    new_plate = init_args.new_plate  # FIXME: re-compute it here
-
-    shutil.copytree(path, new_path)
-
-    print("[maximum_intensity_projection] START")
-    print(f"[maximum_intensity_projection] {path=}")
-    print(f"[maximum_intensity_projection] {new_path=}")
-    print("[maximum_intensity_projection] END")
-
-    out = dict(
-        image_list_updates=[
-            dict(
-                path=new_path,
-                origin=path,
-                attributes=dict(plate=new_plate),
-            )
-        ],
-    )
     return out
